fix(replica): delay adding replica
Adding a replica to an unpublished nexus was causing the child to be
added as online rather than degraded when the volume was subsequently
published. Because of this a rebuild was not run and data corruption
could occur when reading from the newly added child.

This change delays the addition of a replica until a volume is
published; ensuring the child is added as degraded and rebuilt
accordingly.

Also added a new "replica_test" E2E test.